+++ Assets/OwnScripts/Control/Project.cs
@@ -7,7 +7,7 @@ public class Project
     public int Id { get; set; }
     public string Description { get; set; }
     public string Type { get; set; }
-    public Dictionary<string, double> Properties { get; set; }
+    public Dictionary<string, float> Properties { get; set; }
     public Queue<Objective> Objectives { get; set; }
     public Objective CurrentObjective { get; set; }
     // Public List<Problem> Problems {get; set;} -> Generados al azar de entre todos los posibles
@@ -17,14 +17,14 @@ public class Project
         Id = id;
         Description = desc;
         Type = type;
-        Properties = new Dictionary<string, double>();
+        Properties = new Dictionary<string, float>();
         Objectives = new Queue<Objective>();
         CurrentObjective = new Objective();
     }
 
     public Project()
     {
-        Properties = new Dictionary<string, double>();
+        Properties = new Dictionary<string, float>();
         Objectives = new Queue<Objective>();
         CurrentObjective = new Objective();
     }
